:recycle: Large clean-up, move matrix stdlib
