+++ src/lib/matrix_lib.lua
@@ -1,45 +1,6 @@
-function round2(num, numDecimalPlaces)
-	num = num*(math.pow(10,numDecimalPlaces))
-	num = math.floor(num+0.5)
-	return num/(math.pow(10,numDecimalPlaces))
-end
-
-function sleep(n)
-  os.execute("sleep " .. tonumber(n))
-end
-
-function round(x)
-	if(x-math.floor(x)<0.5) then
-		return math.floor(x)
-	else
-		return math.ceil(x)
-	end
-end
-
-function inargs(tosearch)
-	local h = false
-	for i=1,#arg do
-		if(arg[i]==tosearch) then h = true end
-	end
-	return h
-end
-
-weights = {
-	save = function(name,array)
-		local file = io.open(name, "w")
-		file:write(json.encode(array))
-		file:close()
-	end,
-	load = function(name)
-		local file = io.open(name, "r")
-		local json = json.decode(file:read())
-		file:close()
-		return json
-	end	
-}
 
 -- MATRIX LIBRARY, FOR MATRICES
-m = {
+matrix_utilities = {
 	summate = function(array)
 		local total = 0
 		for i=1,#array do
@@ -161,46 +122,4 @@ m = {
 	end
 }
 
-
-function print_r(arr, indentLevel)
-    local str = ""
-    local indentStr = "#"
-
-    if(indentLevel == nil) then
-        print(print_r(arr, 0))
-        return
-    end
-
-    for i = 0, indentLevel do
-        indentStr = indentStr.."\t"
-    end
-
-    for index,value in pairs(arr) do
-        if type(value) == "table" then
-            str = str..indentStr..index..": \n"..print_r(value, (indentLevel + 1))
-        else 
-            str = str..indentStr..index..": "..value.."\n"
-        end
-    end
-    return str
-end
-
-function hr()
-	print("-----------------------------------")
-end
-
-function die ()
-  os.exit(1)
-end
-
-function read_file(file)
-	local f = io.open(file, "r")
-    local content = f:read("*all")
-    f:close()
-    return content
-end
-
-function file_exists(name)
-   local f = io.open(name,"r")
-   if f ~= nil then io.close(f) return true else return false end
-end
+return matrix_utilities
